Snake Game Project app
This repository is made by Krishna Garg, a student at DIT University, Dehradun.

The repository is made to provide the project for Techinical Training. The project is based on the concept of basic Snake Game. The repository contains a ReadMe file for details, a screenshot of the working game and the source code that made it possible.

Krishna Garg Roll no. 210102907